requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
